# warren/openapi.yml
openapi: 3.0.3
info:
  title: warren API
  version: 0.1.0
  description: |
    Coordination hub for LLM agents. Admin (PSK-authenticated) injects
    requests and memos; agents (Bearer-token-authenticated) claim
    requests, respond, and acknowledge memos.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: warren_session
    agentToken:
      type: http
      scheme: bearer
  schemas:
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        class: { type: string }
        type: { type: string, nullable: true }
        model: { type: string }
        authtoken: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, name, class, type, model, authtoken, created_at]
    AgentNew:
      type: object
      properties:
        name: { type: string }
        class: { type: string }
        type: { type: string, nullable: true }
        model: { type: string }
      required: [name, class, model]
    AgentPatch:
      type: object
      properties:
        name: { type: string }
        class: { type: string }
        model: { type: string }
    LoginReq:
      type: object
      properties:
        password: { type: string }
      required: [password]
    LoginRes:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    Request:
      type: object
      properties:
        id: { type: string, format: uuid }
        target_class: { type: string }
        target_type: { type: string, nullable: true }
        payload: {}
        response: { nullable: true }
        status: { type: string, enum: [pending, approved, responded, rejected] }
        claimed_by: { type: string, format: uuid, nullable: true }
        claimed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        responded_at: { type: string, format: date-time, nullable: true }
      required: [id, target_class, target_type, payload, status, created_at]
    RequestNew:
      type: object
      properties:
        target_class: { type: string }
        target_type: { type: string, nullable: true }
        payload: {}
        approved: { type: boolean, default: false }
      required: [target_class, payload]
    RequestRespond:
      type: object
      properties:
        response: {}
      required: [response]
    Memo:
      type: object
      properties:
        id: { type: string, format: uuid }
        target_class: { type: string }
        target_type: { type: string, nullable: true }
        payload: {}
        status: { type: string, enum: [pending, approved, rejected] }
        created_at: { type: string, format: date-time }
      required: [id, target_class, target_type, payload, status, created_at]
    MemoNew:
      type: object
      properties:
        target_class: { type: string }
        target_type: { type: string, nullable: true }
        payload: {}
        approved: { type: boolean, default: false }
      required: [target_class, payload]
...
paths:
  /api/login:
    post:
      summary: Log in as admin (sets warren_session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginReq' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginRes' }
        '401':
          description: Bad PSK
...
  /api/agents:
    get:
      summary: List agents (admin)
      security:
        - adminSession: []
        - agentToken: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Agent' }
    post:
      summary: Create agent (admin); returns server-generated authtoken
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentNew' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
  /api/agents/me:
    get:
      summary: Whoami (agent or admin)
      security:
        - adminSession: []
        - agentToken: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '403':
          description: Admin has no agent identity
  /api/agents/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get agent (admin)
      security: [{ adminSession: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '404': { description: Not found }
    put:
      summary: Patch agent (admin)
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
    delete:
      summary: Delete agent (admin)
      security: [{ adminSession: [] }]
      responses:
        '204': { description: OK }
  /api/requests:
    get:
      summary: List requests (admin); ?status= ?limit= ?offset=
      security: [{ adminSession: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Request' }
    post:
      summary: Create request (admin)
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestNew' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
  /api/requests/incoming:
    get:
      summary: "Agent inbox: approved+unclaimed requests matching your class+type"
      security: [{ agentToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Request' }
  /api/requests/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get request (admin or owning agent)
      security:
        - adminSession: []
        - agentToken: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '403': { description: Forbidden }
  /api/requests/{id}/claim:
    post:
      summary: "Atomically claim a request (agent; must match class+type)"
      security: [{ agentToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Claimed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '409': { description: Not in inbox or already claimed }
  /api/requests/{id}/respond:
    post:
      summary: Respond to your claimed request (agent)
      security: [{ agentToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestRespond' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '409': { description: Not claimed by you }
  /api/requests/{id}/approve:
    post:
      summary: Approve a pending request (admin)
      security: [{ adminSession: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: OK }
  /api/requests/{id}/reject:
    post:
      summary: Reject a pending request (admin)
      security: [{ adminSession: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: OK }
  /api/memos:
    get:
      summary: List memos (admin); ?status= ?limit= ?offset=
      security: [{ adminSession: [] }]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Memo' }
    post:
      summary: Create memo (admin)
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemoNew' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Memo' }
  /api/memos/incoming:
    get:
      summary: "Agent inbox: unacknowledged approved memos matching your class+type"
      security: [{ agentToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Memo' }
  /api/memos/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get memo (admin, or agent if in inbox)
      security:
        - adminSession: []
        - agentToken: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Memo' }
        '403': { description: Forbidden }
  /api/memos/{id}/acknowledge:
    post:
      summary: Acknowledge a memo addressed to you (agent)
      security: [{ agentToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: OK }
        '409': { description: Memo not approved }
  /api/memos/{id}/approve:
    post:
      summary: Approve a pending memo (admin)
      security: [{ adminSession: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: OK }
  /api/memos/{id}/reject:
    post:
      summary: Reject a pending memo (admin)
      security: [{ adminSession: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: OK }